uc2 commit float

+++ src/Main.java
@@ -1,9 +1,8 @@
 import java.util.Scanner;
-//TIP To <b>Run</b> code, press <shortcut actionId="Run"/> or
-// click the <icon src="AllIcons.Actions.Execute"/> icon in the gutter.
+
 class MaxNumber {
-    public int testMax(int num1, int num2, int num3) {
-        int max = num1;
+    public float testMax(float num1, float num2, float num3) {
+        float max = num1;
         if (num2 > max) {
             max = num2;
         }
@@ -13,23 +12,25 @@ class MaxNumber {
         return max;
     }
 }
+
 public class Main {
     public static void main(String[] args) {
 
         Scanner scanner = new Scanner(System.in);
 
         System.out.print("Enter the first number: ");
-        int num1 = scanner.nextInt();
+        float num1 = scanner.nextFloat();
 
         System.out.print("Enter the second number: ");
-        int num2 = scanner.nextInt();
+        float num2 = scanner.nextFloat();
 
         System.out.print("Enter the third number: ");
-        int num3 = scanner.nextInt();
+        float num3 = scanner.nextFloat();
 
         MaxNumber maxObj = new MaxNumber();
-        int maxi=maxObj.testMax(num1, num2, num3);
-        System.out.println("Maximum number is "+maxi);
+        float maxi = maxObj.testMax(num1, num2, num3);
+        System.out.println("Maximum number is " + maxi);
+
         scanner.close();
     }
 }
